[Core] :fire: Fix compilation

+++ mobile/src/main/java/com/mercandalli/android/home/main/MainAdapter.kt
@@ -204,7 +204,7 @@ class MainAdapter constructor(
 
     data class TrainTrafficViewModel(
             val title: String,
-            @TrainManager.Companion.TrainTrafficType val trainTrafficType: Long,
+            @TrainManager.Companion.TrainTrafficType val trainTrafficType: Int,
             val trainTraffic: TrainTraffic?)
     //endregion TrainTraffic
